Fix JobDefinition constructor
